# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LifeLink API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for the LifeLink donation platform API. Primary routes are mounted at root paths and compatibility aliases are also available under /api/v1/*."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Configured server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/admin/profile": {
      "get": {
        "summary": "Get admin profile",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile"
          }
        }
      }
    },
...
    "/admin/system/maintenance": {
      "post": {
        "summary": "Toggle maintenance mode",
        "tags": [
          "Admin - System"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "enabled"
                ],
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Maintenance mode updated"
          }
        }
      },
      "get": {
        "summary": "Get maintenance mode status",
        "tags": [
          "Admin - System"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance status"
          }
        }
      }
    },
    "/admin/statistics": {
      "get": {
        "summary": "Get admin statistics summary",
        "tags": [
          "Admin - Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics summary"
          }
        }
      }
    },
    "/admin/audit-logs": {
      "get": {
        "summary": "List audit logs",
        "tags": [
          "Admin - Audit"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "targetType",
            "schema": {
              "type": "string",
              "enum": [
                "User",
                "Request",
                "Donation",
                "System"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs list"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "List all users with filters and pagination",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "donor",
                "hospital",
                "admin",
                "superadmin"
              ]
            }
          },
          {
            "in": "query",
            "name": "verified",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "in": "query",
            "name": "suspended",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users list"
          }
        }
      }
    },
    "/admin/users/stats": {
      "get": {
        "summary": "Get user statistics",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User statistics"
          }
        }
      }
    },
    "/admin/users/hospital": {
      "post": {
        "summary": "Create a new hospital (admin-created, pre-verified)",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "email",
                  "password",
                  "hospitalName",
                  "hospitalId",
                  "licenseNumber"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "hospitalName": {
                    "type": "string"
                  },
                  "hospitalId": {
                    "type": "number"
                  },
                  "licenseNumber": {
                    "type": "string"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "governorate": {
                        "type": "string"
                      }
                    }
                  },
                  "contactNumber": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hospital created"
          },
          "409": {
            "description": "Email already registered"
          }
        }
      }
    },
    "/admin/users/{id}": {
      "get": {
        "summary": "Get user details by ID",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a user account",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "403": {
            "description": "Cannot delete admin accounts"
          }
        }
      }
    },
    "/admin/users/{id}/verify": {
      "patch": {
        "summary": "Verify a user's email",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User verified"
          }
        }
      }
    },
    "/admin/users/{id}/unverify": {
      "patch": {
        "summary": "Unverify a user's email",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unverified"
          }
        }
      }
    },
    "/admin/users/{id}/suspend": {
      "patch": {
        "summary": "Suspend a user account",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User suspended"
          },
          "403": {
            "description": "Cannot suspend admin accounts"
          }
        }
      }
    },
    "/admin/users/{id}/unsuspend": {
      "patch": {
        "summary": "Unsuspend a user account",
        "tags": [
          "Admin - Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unsuspended"
          }
        }
      }
    },
    "/admin/requests": {
      "get": {
        "summary": "List all blood/organ requests",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "in-progress",
                "completed",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "urgency",
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "critical"
              ]
            }
          },
          {
            "in": "query",
            "name": "bloodType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "blood",
                "organ"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requests list"
          }
        }
      }
    },
    "/admin/requests/stats": {
      "get": {
        "summary": "Get request statistics",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request statistics"
          }
        }
      }
    },
    "/admin/requests/{id}": {
      "get": {
        "summary": "Get full request details with donations",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request details"
          }
        }
      }
    },
    "/admin/requests/{id}/donations": {
      "get": {
        "summary": "Get donations for a specific request",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request donations"
          }
        }
      }
    },
    "/admin/requests/{id}/fulfill": {
      "patch": {
        "summary": "Mark a request as fulfilled",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request fulfilled"
          }
        }
      }
    },
    "/admin/requests/{id}/cancel": {
      "patch": {
        "summary": "Cancel a request",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request cancelled"
          }
        }
      }
    },
    "/admin/requests/{id}/broadcast": {
      "post": {
        "summary": "Broadcast request to eligible donors",
        "tags": [
          "Admin - Requests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Broadcast sent"
          }
        }
      }
    },
    "/admin/analytics/dashboard": {
      "get": {
        "summary": "Dashboard summary with key metrics",
        "tags": [
          "Admin - Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard summary"
          }
        }
      }
    },
...
    "/admin/analytics/blood-types": {
      "get": {
        "summary": "Blood type distribution",
        "tags": [
          "Admin - Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Blood type distribution"
          }
        }
      }
    },
...
  },
...
}